Move electron-app.js to main/index.js. We can speed up the build later by not running main through babel and browserify, as it's not necessary under the node environment (would just need to change import to require).
