+++ resource-compile/package.json
@@ -1,7 +1,7 @@
 {
   "name": "james",
   "description": "HTTP/HTTPS proxy built in Electron and React",
-  "main": "electron-app.js",
+  "main": "main/index.js",
   "author": "Julian Hollmann <hollmann@uxebu.com>",
   "license": "MIT",
   "repository": {
